+++ folioreader/src/main/java/com/folioreader/ui/folio/mediaoverlay/MediaControllerCallbacks.java
@@ -1,0 +1,14 @@
+package com.folioreader.ui.folio.mediaoverlay;
+
+/**
+ * Created by gautam on 21/6/17.
+ */
+
+public interface MediaControllerCallbacks {
+
+    void highLightText(String text);
+
+    void highLightTTS();
+
+    void resetCurrentIndex();
+}
